pyproject.toml
[tool.poetry]
name = "topical"
version = "0.1.0"
description = ""
authors = ["John Giorgi <johnmgiorgi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.26.0"
guidance = "^0.0.64"
nltk = "^3.8.1"
biopython = "^1.81"
instructorembedding = "^1.0.1"
sentence-transformers = "^2.2.2"
diskcache = "^5.6.3"

[tool.poetry.group.dev.dependencies]
watchdog = "^3.0.0"
isort = "^5.12.0"
black = "^23.9.1"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pytest = "^7.4.2"

[tool.black]
# This configuration is adapted from: 
# - https://github.com/allenai/allennlp/blob/main/pyproject.toml
# - https://github.com/allenai/cached_path/blob/main/pyproject.toml
# - https://github.com/huggingface/transformers/blob/main/pyproject.toml
line-length = 119

include = '\.pyi?$'

exclude = '''
(
      __pycache__
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.vscode
    | \.venv
    | \bdist\b
    | \bdoc\b
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
